design tela listar, botões, topbar

+++ src/app/core/tasks/components/list/list.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { FormBuilder, FormGroup } from '@angular/forms';
 import { Router } from '@angular/router';
+
+import { Task } from 'src/app/shared/model/task.model';
 import { TaskService } from '../../services/task.service';
 
 @Component({
@@ -10,7 +12,7 @@ import { TaskService } from '../../services/task.service';
 })
 export class ListComponent implements OnInit {
 
-  tasks: any;
+  tasks: Task[];
   form: FormGroup;
 
   constructor(
@@ -27,34 +29,46 @@ export class ListComponent implements OnInit {
   formInit() {
     this.form = this.formBuilder.group({
       titulo: [null],
-      descricao: [null]
+      descricao: [null],
+      concluida: false
     });
   }
 
   listTasks() {
-    this.taskService.getTasks().subscribe((data) => {
+    this.taskService.getTasks().subscribe((data: Task[]) => {
       this.tasks = data;
     });
   }
 
   setTask() {
-    let task = this.form.value;
+    let task: Task = this.form.value;
     this.taskService.postTask(task).subscribe(() => {
       this.tasks.push(task);
     });
   };
 
   deleteTask(id) {
-    this.taskService.deleteTask(id).subscribe(() => {
-      const index = this.tasks.findIndex(task => task.id === id);
-      if(index>=0) {
-        this.tasks.splice(index, 1);
-      }
-    });
+    if (window.confirm('Deseja deletar essa tarefa ?')) {
+
+      this.taskService.deleteTask(id).subscribe(() => {
+        const index = this.tasks.findIndex((task: Task) => task.id === id);
+        if (index >= 0) {
+          this.tasks.splice(index, 1);
+        }
+      });
+    }
+  }
+
+  editTask(task: Task) {
+    task.concluida = !task.concluida;
+    this.taskService.putTask(task).subscribe(() => {
+      this.listTasks();
+    })
   }
 
   onSubmit() {
     this.setTask();
+    this.form.reset();
   }
 
 }
